Fixes application initialization to run rspec.
